ingestion/functions/pyproject.toml
[tool.poetry]
name = "ingestion functions"
version = "0.1.0"
description = "Ingestion functions for Global.health"
authors = ["Global.health: a Data Science Initiative <info@global.health>"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "1.13.26"
google = "2.0.3"
google-auth = "1.16.1"
requests-mock = "1.8.0"
pycountry = "20.7.3"
pytest = "^6.2.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
